Containment added, declarations methods changed
I changed the method format to include the description and the class
they are contained, because methods that had the same name were
eliminated.

Issue: I am trying to figure out is how to identify interfaces. Right
now they are identified in implements but not in declaration.

I might try to overlook it to make overrides work.