MethodOverrides
Issues:
Still counts the init we can easily eliminate this but I do not know if
we all agree on that so I left it that way.

Also until I find or we find a way to recognize interfaces in
declarations I label them as classes everywhere so it is consistent.